cortexm: Halt the device before reading the ROMTABLE
Many devices need to be halted to read the romtable. As this was not done
before, the cortexm_forced_halt workaround was introduced.
We now halt the device before reading the ROMTABLE. STM32F7 also needs to be
enabled for debugging in sleep. But as a ADIV5_DPv2 device, we can already
detect it.